Modify: 리팩토링
