+++ src/pages/Hyunjung/Main/FeedsArticle/FeedsArticle.scss
@@ -1,0 +1,11 @@
+.feeds-article {
+  padding: 0 10px;
+
+  .feeds-article-id {
+    font-weight: 700;
+  }
+
+  .feeds-article-over {
+    color: #cacaca;
+  }
+}
