+++ src/pages/Hyunjung/Main/Header/Header.js
@@ -1,12 +1,12 @@
 import React from 'react';
+import './Header.scss';
 
-const Header = ({ feedsArr }) => {
-  // console.log(feedsArr.HeaderUrl);
+const Header = ({ header }) => {
   return (
     <div className="header">
       <div className="header-id">
-        <img alt="profile-img" src={feedsArr.headerUrl} className="id-img" />
-        <div className="id">{feedsArr.headerName}</div>
+        <img alt="profile-img" src={header.headerUrl} className="id-img" />
+        <div className="id">{header.headerName}</div>
       </div>
       <div className="header-dot">
         <i className="fa-solid fa-ellipsis" />
