+++ src/pages/Hyunjung/Main/HyunjungMain.js
@@ -1,25 +1,15 @@
 import React, { useEffect, useState } from 'react';
 import './HyunjungMain.scss';
 import Nav from '../../../components/Nav/Nav';
-import Comment from './Comment/HyunjungComment';
 import Header from './Header/Header';
-import HyunJungAside from './Aside/HyunjungAside';
+import Aside from './Aside/Aside';
 import FeedsIcon from './FeedsIcon/FeedsIcon';
 import FeedsArticle from './FeedsArticle/FeedsArticle';
 import CommentList from './CommentList/CommentList';
 
 const HyunJungMain = props => {
-  // const [commentList, setCommentList] = useState([]);
-
   const [feedsArr, setFeedsArr] = useState([]);
 
-  // useEffect(() => {
-  //   fetch('http://localhost:3000/data/commentData.json', { method: 'GET' })
-  //     .then(res => res.json())
-  //     .then(data => {
-  //       setCommentArr(data);
-  //     });
-  // }, []);
   useEffect(() => {
     fetch('http://localhost:3000/data/commentData.json', { method: 'GET' })
       .then(res => res.json())
@@ -37,7 +27,7 @@ const HyunJungMain = props => {
           {feedsArr.map(feed => (
             <section className="left" key={feed.feedId}>
               <div className="feeds">
-                <Header feedsArr={feed} />
+                <Header header={feed} />
                 <div className="feeds-pic">
                   <img alt="main-pic" src={feed.imgUrl} />
                 </div>
@@ -46,32 +36,12 @@ const HyunJungMain = props => {
                 <div>
                   <FeedsArticle feedsArr={feed} />
                   <CommentList feedsArr={feed} />
-                  {/* <div>
-                    <Comment feedsArr={feed.comment} newComment={commentArr} />
-                    <div className="comment-time">55분 전</div>
-                    <form
-                      className="feeds-comment-input"
-                      onSubmit={onFormSubmit}
-                    >
-                      <div className="comment-input">
-                        <input
-                          className="comment-text"
-                          placeholder="댓글 쓰기..."
-                          onChange={onInputChange}
-                          value={inputValue}
-                        />
-                      </div>
-                      <div className="comment-button">
-                        <button className="comment-btn">게시</button>
-                      </div>
-                    </form>
-                  </div> */}
                 </div>
               </div>
             </section>
           ))}
         </article>
-        <HyunJungAside />
+        <Aside />
       </main>
     </>
   );
