+++ src/pages/Hyunjung/Main/HyunjungMain.scss
@@ -15,221 +15,15 @@
       width: 100%;
       border: 1px solid #cacaca;
 
-      .header {
-        display: flex;
-        justify-content: space-between;
-        height: 60px;
-        width: 100%;
-        border-bottom: 1px solid #cacaca;
-
-        .header-id {
-          display: flex;
-          justify-content: flex-start;
-          align-items: center;
-          padding: 14px 4px 14px 16px;
-
-          .id-img {
-            width: 32px;
-            height: 32px;
-            border-radius: 25px;
-          }
-          .id {
-            margin-top: 5px;
-            padding: 0px 10px;
-          }
-        }
-        .header-dot {
-          padding: 20px 25px;
-        }
-      }
       .feeds-pic {
         img {
           width: 100%;
         }
       }
-      .feeds-icon {
-        display: flex;
-        justify-content: space-between;
-        padding: 10px 16px;
 
-        .fa-icon {
-          padding-right: 10px;
-          font-size: 1em;
-        }
-        .feeds-icon-left {
-          span {
-            .feeds-icon-heart {
-              margin-left: -8px;
-            }
-
-            .feeds-icon-bookmark {
-              margin-left: auto;
-            }
-          }
-        }
-      }
       .feeds-like {
         padding: 10px 10px;
         text-align: initial;
-      }
-      div {
-        .feeds-article {
-          padding: 0 10px;
-          text-align: initial;
-          .feed-article-id {
-            font-weight: 50px;
-          }
-
-          .feeds-article-over {
-            color: #cacaca;
-          }
-        }
-        div {
-          .feeds-comment {
-            padding: 10px;
-            //li
-          }
-
-          .comment-time {
-            padding-left: 10px;
-            color: #cacaca;
-            text-align: initial;
-          }
-          .feeds-comment-input {
-            display: flex;
-            // padding-top: 10px;
-            padding: 10px;
-            border-top: 1px solid #cacaca;
-            .comment-input {
-              width: 90%;
-              outline: 0;
-              border: 0;
-              text-align: left;
-              .comment-text {
-                height: 18px;
-                padding: 10px;
-                background-color: #fafafa;
-                outline: 0;
-                border: 0;
-              }
-            }
-            .comment-button {
-              display: flex;
-              .comment-btn {
-                background-color: #fafafa;
-                color: #0095f6;
-                outline: 0;
-                border: 0;
-                font-size: 1em;
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
-.right {
-  .main-right {
-    width: 100%;
-    padding-left: 30px;
-    box-sizing: border-box;
-    div {
-      .aside-header {
-        display: flex;
-        padding: 10px;
-
-        .header-aside-img {
-          height: 40px;
-          width: 40px;
-          outline: 0;
-          border: 0;
-          border-radius: 20px;
-        }
-        .aside-id {
-          flex-direction: column;
-          padding-left: 10px;
-          .aside-id-bottom,
-          .aside-id-time {
-            color: #cacaca;
-          }
-        }
-      }
-      .aside-form {
-        .recommend,
-        .story {
-          margin-top: 15px;
-          border: 1px solid #cacaca;
-          background-color: #ffffff;
-          .story-header {
-            display: flex;
-            justify-content: space-between;
-            padding: 10px;
-            .story-title {
-              color: #cacaca;
-            }
-          }
-          .aside-header {
-            .aside-img {
-              height: 35px;
-              width: 35px;
-              border: 2px solid red;
-              border-radius: 20px;
-              padding: 2px;
-            }
-          }
-          .aside-id {
-            flex-direction: column;
-            padding-left: 10px;
-            .aside-id-bottom,
-            .aside-id-time {
-              color: #cacaca;
-            }
-          }
-          .aside-right {
-            padding-top: 10px;
-            color: #cacaca;
-          }
-
-          .recommend-header {
-            display: flex;
-            justify-content: space-between;
-            padding: 10px;
-            .recommend-title {
-              color: #cacaca;
-            }
-          }
-          .member {
-            display: flex;
-            justify-content: space-between;
-            padding: 10px;
-            .member-header {
-              display: flex;
-              .aside-img {
-                height: 35px;
-                width: 35px;
-                border: 2px solid red;
-                border-radius: 20px;
-                padding: 2px;
-              }
-            }
-            .follow {
-              display: flex;
-              align-items: center;
-              color: #0095f6;
-            }
-          }
-        }
-        div {
-          .aside-info {
-            padding-top: 10px;
-            color: #cacaca;
-          }
-        }
-        .aside-right {
-          padding-top: 10px;
-          color: #cacaca;
-        }
       }
     }
   }
@@ -239,10 +33,6 @@
   .main-feeds {
     .left {
       padding: 0 15px;
-    }
-
-    .right {
-      display: none;
     }
   }
 }
